Fix/logtime page

+++ database/migrations/2025_10_20_032509_create_logtimes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('task_id')->constrained()->onDelete('cascade');
             $table->dateTime('date');
-            $table->integer('time_used');
+            $table->decimal('time_used');
             $table->timestamps();
         });
     }
